Dynamic string and static string integration
